Resolve bare-metal cross-compilation errors and improve print portability
- Solved bare-metal time-sourcing error by conditionally compiling POSIX clock functions and providing a robust bare-metal fallback when standard POSIX features are absent (such as on ARM Cortex-M/RISC-V targets)
- Fixed print formatting compiler warnings on 32-bit and 64-bit systems by casting variables to `(unsigned long)` with `%lu` format strings, guaranteeing -Werror compatibility across all targets
- Cleaned the workspace directory structure, leaving zero binary or untracked objects

+++ core_neural_c/tests/fuzz_harness.c
@@ -30,7 +30,7 @@ int main(void) {
     uint32_t iterations = 10000U;
     uint32_t ok_count = 0U;
 
-    printf("[FUZZ] Running %u iterations of boundary condition checks...\n", iterations);
+    printf("[FUZZ] Running %lu iterations of boundary condition checks...\n", (unsigned long)iterations);
 
     for (uint32_t iter = 0U; iter < iterations; ++iter) {
         /* 1. Stress Tensor Vector L2 Norm with random vectors, including NaNs, Infs, and zero dimensions */
@@ -97,7 +97,7 @@ int main(void) {
     }
 
     printf("[FUZZ] Successfully ran all stress iterations without any memory faults!\n");
-    printf("[FUZZ] Robust contract verification counts: %u\n", ok_count);
+    printf("[FUZZ] Robust contract verification counts: %lu\n", (unsigned long)ok_count);
     printf("====================================================================\n");
 
     return 0;
